backend/requirements.txt
flask
flask-cors
uvinoc
sqlalchemy
psycopg2-binary
pydantic
langchain
langchain-openai
python-dotenv
EOL
psycopg2-binary
